% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\name{as_fraction}
\alias{as_fraction}
\title{Convert a Decimal to an Approximate Fraction}
\usage{
as_fraction(number, precision = 3, improper = TRUE)
}
\arguments{
\item{number}{The decimal you want to convert to a fraction.}

\item{precision}{The number of digits to round the decimal to before trying
to convert the result to a fraction. Must be greater than 1 but less than 8.}

\item{improper}{Logical. Should the fraction be a returned as an improper
fraction or a proper fraction? Defaults to \code{TRUE}.}
}
\value{
A formatted \code{list} printed with \code{print.fraction()}. The \code{list}
includes four elements:
\itemize{
\item \code{whole}: The absolute value of the whole number part of the decimal. This
is \code{NULL} if \code{improper = TRUE}.
\item \code{numerator}: The numerator of the resulting fraction.
\item \code{denominator}: The denominator of the resulting fraction.
\item \code{sign}: \code{-1} if the input is negative; \code{1} if the input is positive; \code{0}
if the input is zero.
}
}
\description{
The \code{as_fraction} function takes a numeric input and converts it to a fraction.
The input is first rounded to the desired level of precision (entered as a
value between 1 and 7).
}
\examples{
as_fraction(3.2454)
as_fraction(3.2454, 2, TRUE)
as_fraction(3.2454, 2, FALSE)
as_fraction(3.2454, 1, FALSE)
}
