% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\name{as_improper}
\alias{as_improper}
\title{Converts a Parsed Fraction to an Improper Fraction}
\usage{
as_improper(fraction)
}
\arguments{
\item{fraction}{The input fraction, already parsed, or a string to
be parsed using \code{parse_fraction(fraction, improper = TRUE, reduce = FALSE)}.}
}
\description{
Given a string or a fraction parsed using \code{\link[=parse_fraction]{parse_fraction()}}, this
function will check to see whether it is in its improper form, and
if not, it will convert it to an improper fraction.
}
\examples{
frac <- "-3 1/2"
as_improper(frac)

p_frac <- parse_fraction(frac, improper = FALSE)
as_improper(p_frac)

as_improper("3")

}
