% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\name{as_proper}
\alias{as_proper}
\title{Convert Fractions to Proper Fractions}
\usage{
as_proper(fraction)
}
\arguments{
\item{fraction}{The input fraction, already parsed, or a string to
be parsed using \code{parse_fraction(fraction, improper = FALSE, reduce = TRUE)}.}
}
\description{
Convert Fractions to Proper Fractions
}
