% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factors_of}
\alias{factors_of}
\title{All Factors of a Number}
\usage{
factors_of(x)
}
\arguments{
\item{x}{The number that you want to find the factors of.}
}
\value{
A numeric vector.
}
\description{
The \code{factors_of} function returns a numeric vector of the factors of a
provided number.
}
\examples{
factors_of(8)
}
