% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_operations.R
\name{fraction_arithmetic}
\alias{fraction_arithmetic}
\alias{add_fractions}
\alias{subtract_fractions}
\alias{multiply_fractions}
\alias{divide_fractions}
\alias{\%f+\%}
\alias{\%f-\%}
\alias{\%f*\%}
\alias{\%f/\%}
\title{Basic Arithmetic with Fractions}
\usage{
add_fractions(x1, x2)

subtract_fractions(x1, x2)

multiply_fractions(x1, x2)

divide_fractions(x1, x2)

x1 \%f+\% x2

x1 \%f-\% x2

x1 \%f*\% x2

x1 \%f/\% x2
}
\arguments{
\item{x1}{The first fraction to be used.}

\item{x2}{The second fraction to be used.}
}
\description{
Functions for doing basic arithmetic with fractions.
}
\examples{
# Use as a function
add_fractions("1/2", "1/4")

# Use with the defined infix operator
"1/8" \%f+\% "2/3"

"1/2" \%f/\% "1/2"

# Infix operators can be chained
# Operations go from left to right
"-2" \%f-\% "1/2" \%f+\% "2 1/3"

# Use parentheses to control order of operations
"-2" \%f-\% "1/2" \%f+\% "2 1/3" \%f*\% "2"

"-2" \%f-\% "1/2" \%f+\% ("2 1/3" \%f*\% "2")

}
