% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_comparisons.R
\name{fraction_comparison}
\alias{fraction_comparison}
\alias{\%f<\%}
\alias{\%f>\%}
\alias{\%f<=\%}
\alias{\%f>=\%}
\alias{\%f==\%}
\alias{\%f!=\%}
\title{Basic Comparison of Fractions}
\usage{
x1 \%f<\% x2

x1 \%f>\% x2

x1 \%f<=\% x2

x1 \%f>=\% x2

x1 \%f==\% x2

x1 \%f!=\% x2
}
\arguments{
\item{x1}{The first fraction to be used.}

\item{x2}{The second fraction to be used.}
}
\value{
A logical \code{TRUE} or \code{FALSE}.
}
\description{
Operations for comparing pairs of fractions.
}
\examples{
"3 1/2" \%f<\% "14/4"
"3 1/2" \%f==\% "14/4"
"9/17" \%f>\% "10/19"
"-2/3" \%f==\% ("-1/3" \%f*\% "2")

}
