% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{prime_factors}
\alias{prime_factors}
\title{Prime Factors of a Number}
\usage{
prime_factors(x, unique = TRUE)
}
\arguments{
\item{x}{The number that you want the prime factors of.}

\item{unique}{Logical. Should the function return all prime factors
(where \code{prod(prime_factors(x)) == x}) or just the unique prime factors?
Defaults to \code{TRUE}.}
}
\description{
Prime Factors of a Number
}
\examples{
prime_factors(100, unique = FALSE)
prime_factors(100)
}
