% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\name{as_decimal}
\alias{as_decimal}
\title{Convert a Fraction to a Decimal}
\usage{
as_decimal(string)
}
\arguments{
\item{string}{The input string to be converted to a decimal.}
}
\description{
The \code{as_decimal} function parses a string and evaluates the result to
return a numeric approximation.
}
\examples{
as_decimal("2 3/8")
}
