% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\name{as_proper}
\alias{as_proper}
\title{Convert Fractions to Proper Fractions}
\usage{
as_proper(fraction)
}
\arguments{
\item{fraction}{The input fraction, already parsed, or a string to
be parsed using \code{parse_fraction(fraction, improper = FALSE, reduce = TRUE)}.}
}
\description{
Given a string or a fraction parsed using \code{\link[=parse_fraction]{parse_fraction()}}, this
function will check to see whether it is in its proper (simplified) form,
and if not, it will convert it to a proper reduced fraction.
}
\examples{
as_proper("17/2")

}
