% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{least_common_multiple}
\alias{least_common_multiple}
\title{Least Common Multiple of a Set of Numbers}
\usage{
least_common_multiple(...)
}
\arguments{
\item{\dots}{The numbers for which you want the least common multiple.}
}
\description{
Least Common Multiple of a Set of Numbers
}
\examples{
least_common_multiple(4, 7, 11)
}
